Day 3 Solved
Part 1: Identify the item type that appears in both compartments of each rucksack and determine the sum of the priorities of those item types.
Part 2: Identify the item type that corresponds to the badges of each group of three Elves and determine the sum of the priorities of those item types.