Day 7 Solved
Part 1: Using a terminal-like input build a tree representation of file system and fine the sum of size of directories that are less than or equal to size 100000.
Part 2:  Find the size of the smallest directory that, if deleted, would free up enough space on the filesystem to run the update.