Day 8 Solved
Part 1. Determine the number of "visible" positions in a grid of digits. A position is considered visible if its value is the highest compared to the surrounding elements in four subarrays: to its left and right in the row, and above and below in its column.
Part 2. Identify the maximum scenic score within the grid. The scenic score of a position is calculated as the product of the counts of visible positions to its left, right, top, and bottom.