Day 9 Solved
Part 1: Simulate the movement of a rope's tail on a two-dimensional grid as its head follows a series of given motions. Determine and count the number of unique positions the tail visits on this grid.
Part 2: Extend the earlier rope simulation to account for a longer rope that consists of ten knots. Simulate the movement of this longer rope's tail on a two-dimensional grid as its head follows a series of given motions. Determine and count the number of unique positions the tail (the 9th knot, since we're counting from the head as 0) visits on this grid.