2023 Day 2 Solved
Part 1 involved determining which games (from a series of games where an Elf reveals sets of colored cubes) could have been possible given a fixed number of cubes (12 red, 13 green, 14 blue). The task was to find games where the maximum number of any color revealed at once did not exceed the available cubes and then sum the IDs of these possible games.
Part 2 required calculating the minimum number of each color of cubes needed to make each game possible. This involved identifying the highest count of each color shown in each game. The 'power' of each game's minimum set of cubes was calculated by multiplying the counts of red, green, and blue cubes. The task was to find the sum of these powers for all games.