2023 Day 5 Solved
Part 1 required finding the lowest location number from a given list of seeds. The given mapping system converts ranges of numbers from one category to the next, culminating in the location category.

Part 2 expanded the problem by interpreting the initial seed numbers as ranges, which significantly increased the number of seeds to be processed.