2023 Day 11 Solved
Part 1 required calculating the sum of the shortest paths between every pair of '#' symbols in a grid of ASCII characters. The challenge was that every row and column without a '#' symbol required two steps to cross.

Part 2 presented the same task, but with a twist: rows and columns without '#' symbols took a million steps to cross.